+++ ricaun.Revit.DA.Example/Revit/App.cs
@@ -1,12 +1,13 @@
 using Autodesk.Revit.ApplicationServices;
 using Autodesk.Revit.DB;
+using ricaun.Revit.DA.Extensions;
 using System;
 
 namespace ricaun.Revit.DA.Example.Revit
 {
     public class App : DesignApplication<DesignAutomation>
     {
-
+        public override bool UseConsoleLog => true;
     }
 
     public class DesignAutomation : IDesignAutomation
@@ -15,6 +16,9 @@ namespace ricaun.Revit.DA.Example.Revit
         {
             var output = RevitOutputModel.Create(application);
             output.Save();
+
+            Console.WriteLine(output.ToJson());
+
             return true;
         }
     }