+++ ricaun.Revit.DA.Tests/Tests.cs
@@ -11,7 +11,7 @@ namespace ricaun.Revit.DA.Tests
     {
         //[TestCase("2024", "4.7", "19", null)]
         //[TestCase("2025", "4.7", "19", null)]
-        [TestCase("2019", "4.7", "19", "Example")]
+        //[TestCase("2019", "4.7", "19", "Example")]
         //[TestCase("2020", "4.7", "19", "Example")]
         [TestCase("2024", "4.8", "21", "Example")]
         [TestCase("2025", "8.0", "25", "Example")]