+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Initial release.
 ### DesignApplication
 - Add `DesignApplication<T>` option.
+- Add `UseConsoleLog` to enable console log for the internal `DesignApplication` methods.
 - Update `DesignApplication` startup and shutdown methods to virtual.
 - Add `DesignApplicationLoader` to fix [Design Automation for Revit ignores PackageContents.xml configuration.](https://github.com/ricaun-io/RevitAddin.DA.Tester/issues/7)
 - Add `ExternalServer` to fix [Design Automation for Revit ActiveAddInId is null](https://github.com/ricaun-io/RevitAddin.DA.Tester/issues/9)